src/tests/requirements.txt
openai
pytest
pytest-asyncio
pytest-watch
pytest-httpserver
azure-ai-formrecognizer
